backend/requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
sqlalchemy==2.0.44
alembic==1.17.2
PyMySQL==1.1.0
cryptography==41.0.7
pydantic==2.12.5
openai==2.9.0
tenacity==9.1.2
hypothesis==6.148.7
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-multipart==0.0.20
gunicorn==21.2.0
